requirements.txt
asgiref==3.5.2
attrs==22.1.0
coverage==6.5.0
Django==4.1.4
djangorestframework==3.14.0
exceptiongroup==1.0.4
flake8==6.0.0
iniconfig==1.1.1
mccabe==0.7.0
model-bakery==1.9.0
packaging==22.0
pluggy==1.0.0
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.0
pytest-cov==4.0.0
pytest-django==4.5.2
pytz==2022.6
sqlparse==0.4.3
tomli==2.0.1
